Ignore the test-ecs project
This is causing the script to fail with a permissions error.  The
project is in the `govuk` organisation, but the API token doesn't have
permission to read its stats.

This is a test project belonging to the replatforming team and it's
not useful to report its statistics to statsd, so skip it.